# frontend/src/lib/api/swagger3.json
{
    "openapi": "3.0.0",
...
    "info": {
        "description": "API системы мониторинга репутации брендов BrandRadar",
        "title": "BrandRadar API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/sources": {
            "get": {
                "description": "Возвращает список источников с пагинацией",
                "tags": [
                    "Sources"
                ],
                "summary": "Список источников",
                "parameters": [
                    {
                        "description": "Фильтр по UUID бренда",
                        "name": "brand_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                                                        },
                                                        {
                                                            "type": "object",
                                                            "properties": {
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SourceResponse"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новый источник данных (telegram, web, rss)",
                "tags": [
                    "Sources"
                ],
                "summary": "Создать источник",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.CreateSourceRequest"
                            }
                        }
                    },
                    "description": "Данные источника",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SourceResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sources/{id}": {
            "get": {
                "description": "Возвращает источник по ID",
                "tags": [
                    "Sources"
                ],
                "summary": "Получить источник",
                "parameters": [
                    {
                        "description": "UUID источника",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SourceResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет источник по ID",
                "tags": [
                    "Sources"
                ],
                "summary": "Удалить источник",
                "parameters": [
                    {
                        "description": "UUID источника",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sources/{id}/toggle": {
            "post": {
                "description": "Переключает active/inactive статус источника",
                "tags": [
                    "Sources"
                ],
                "summary": "Переключить статус источника",
                "parameters": [
                    {
                        "description": "UUID источника",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SourceResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.CreateSourceRequest": {
                "type": "object",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "config": {
                        "type": "string",
                        "example": "{\"chat\":\"@brand_news\"}"
                    },
                    "name": {
                        "type": "string",
                        "example": "News channel"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "web",
                            "rss",
                            "telegram"
                        ],
                        "example": "telegram"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "message": {
                        "type": "string",
                        "example": "resource not found"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData": {
                "type": "object",
                "properties": {
                    "items": {},
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                    },
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.SourceResponse": {
                "type": "object",
                "properties": {
                    "config": {
                        "type": "string",
                        "example": "{\"chat\":\"@brand_news\"}"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "News channel"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "type": {
                        "type": "string",
                        "example": "telegram"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
        }
    }
}